Refactor: Move model info retrieval based on model name from build workflow to LLM node, CrewAI node, and Classifier node
